Remove Duplicate Instances
This commit removes duplicate instances from being written to source 
files. This is done during the sorting phase since the insertion sort
is being used and already goes through each instance.

The criteria for a duplicate instance is an instance with the same file
name, byte offset and length. This should be sufficient in determining
duplicate instances as there can only be one instance at a specific 
location in a file. Having a different id or different name may be 
possible, so they are not used in determining whether an instance is a
duplicate.

Issues: #4