Add DiffConfig Object
This commit adds a diff config object that can be used to configure
the diff linker. This config object can be created and passed in the
constructor to manage the various options of the diff linker.

Eventually this object will also be available for the diff mapper
since the two modules are very similar.