Use DiffConfig Object
This commit changes the DiffLinker to use the DiffConfig object
instead of storing config variables within the object itself.

The DiffConfig object also follows a Builder paradigm for creating
instances. A DiffConfigBuilder class was defined to support setter
chaining for parameters. This allows a consumer to construct a
DiffConfig object by using the builder, chaining parameters setters,
then calling the build function.