Add DiffConfig Base Constructor
The current implementation of the DiffConfigBuilder does not easily
allow a consumer to save a base set of config options, but change
a couple of options between two configs. A new static builder
function allows a consumer to create a new instance of the base
builder, chaining the option changes necessary before building
the new config.