Add Config Builder from JSON file
This commit adds an option to build a config based on an input
json file. The config will try to parse the given file as a json
object and if it fails will simply return a default config builder.

If a json object is parsed, it will look for values based on given
keys. If a null value is found for a key, the default value will be
used by the config builder. If a non null value is found, that value
will be used in place of the default when returning the builder

This commit also changes the default values to use relative paths
based on the current working directory. The directory structures
matches a proposed structure of the project so ensure consistency.