Add Config File Parser Unit Tests
This commit adds Unit tests for the DiffConfig file builder. Various
partial config files are written to temporary files and are checked
to make sure the right values are different from the default.

This commit also fixes an implementation bug due to invalid casting
from Object to String when parsing the types key.