Move Shared Files to DiffMapper
This commit moves the shared DiffConfig and Utils files over from the
DiffLinker project to the DiffMapper project. This is a temporary
measure to easily get the definitions to this project. Eventually,
it would be better to break out the shared codebase into a separate
project so it can be developed independently.