Start Adding Command Line Args
This commit begins adding command line arg support using the apache
commons-cli library. This will allow the built jar to load a
configuration file from the command line instead of having to
hardcode and rebuild the project.

This commit also adjusts pom.xml to add several more steps to the
package rule. This includes copying dependencies over to the build
directory and building the jar referencing the lib folder, as well
as assemblying all classes along with dependency jar together