Add Testing Resources
This commit adds the necessary resources to run the unit tests for the
InstanceTracker. This was previously missed because the directories were
empty. They should be empty as we parse the directory names to get info
on each instance, but git does not allow empty directories to be tracked.
Therefore, .keep files have been added to each leaf directory so git tracks
them properly.