Remove Unnecessary Catch Blocks
This commit removes the unnecessary catch blocks in DiffMapper when
executing git setup and cleanup commands. They only caught exceptions
and did nothing, which meant that if an error occured there would be
no way to determine what caused the error.

Instead we now throw and exception, then catch it in the setup and
cleanup methods. This allows us to print the message from the
exception, then throw a new exception up the call stack so consumers
can deal with the error