Move DiffMapper runs to Temporary Folder
This commit moves the tests for DiffMapper into a separate folder so
they don't interfere with the structure of this repository. A temporary
folder is created and then files are copied over to it. All operations
then happen in the test folder so they are automatically cleaned up
when the test ends.

This commit also fixes an issue with the DiffMapper that is caused by
an assumption of the input config kernel dir path. Now the directory
does not need to include a '/' at the end of it.