Add InstanceParserManager
This commit adds a manager for InstanceParsers so new ones can be added.
This allows the ability to create an InstanceParser subclass, then have
it added to a shared list of Parsers.

The InstanceTracker can now get all the parsers dynamically and try to
parse each entry using the specified format.

Issue: #60