Use InstanceParsers in InstanceTracker
This commit updates the InstanceTracker to use InstanceParsers to parse
entries. It also adds several overloaded methods for more options when
trying to parse instances. This allows the ability to control parsing
a specific entry and whether or not to default to trying all parsers if
the specified one fails.

Issue: #60