Add the ability to filter instances by type
This commit adds the ability to filter by type when tracking instances.
The assumption is that underneath the results directory, there is a
directory for each instance type. The filtering works by comparing these
directories to the list provided through a constructor. This change is
a non-breaking change because the original constructor remains and sets
the types list to include the default values of mutex and spin.

The version of shared utils is also updated to 1.1.0 to signify this
change.

Issue: #58