Fix Invalid Offset Bug
This commit fixes an edge case with the DiffMapper where the offset for
the last instance tries to pull more data than is available. This has
been fixed by using the read() method instead of readFully() and keeping
track of the number of bytes read. This will be 40 for most of the file
and only be less than 40 when we reach the end of the file.