Implement Rebase Recovery Strategy
This commit implements the rebase recovery strategy, by doing an add and
continue of the rebase. This completes when either the rebase is done,
or we have reached a maximum number of attempts.

The other git commands have also been modified to use the new execute
method from Utils.

Issue: #61