Fix Tests for Windows
This commit fixes tests so they pass locally on Windows.

In the DiffLinker Project, tests were failing due to wrong copy commands (cp vs robocopy). This has been fixed along with a bug due to the CRLF encoding for files. The files are rewritten by the DiffLinker before searching by offset to fix this problem.

In the DiffMapper Project, tests were failing due to wrong copy commands (cp vs robocopy). Several commands were added to the cleanup to make sure the DiffMapper state doesn't get messed up due to existing branches. We also tag the state of the diffMapper with a "<new_tag>-mapped" so once a map has been created, it remains in git history.

In RssReader, the kernel feed was wrong due to Windows URL parsing. This was fixed.

In Shared, the OS name is checked to make sure we prepend the command with a "cmd /c" in order to invoke the command prompt. This allows Windows Commands to be run with cmd.